[chore] RNSN-51 패키지 명 변경
