+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.application.name=demo
+spring.application.name=smore
 
 # DataSource settings: set here your own configurations for the mysql database connection.
 spring.datasource.url=jdbc:mysql://3.35.47.125:3306/smore_DB?useSSL=false&serverTimezone=UTC
