+++ src/test/java/com/als/SMore/ApplicationTests.java
@@ -1,10 +1,10 @@
-package com.als.SMore.demo;
+package com.als.SMore;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class DemoApplicationTests {
+class ApplicationTests {
 
 	@Test
 	void contextLoads() {
